tools/course-average-fetcher/pyproject.toml
[project]
name = "course-average-fetcher"
version = "0.0.0"
description = "A tool to fetch latest course average information from a crowdsourced spreadsheet"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "arrg>=0.1.0",
  "google-api-python-client>=2.164.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.1",
]

[tool.ruff]
indent-width = 2
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 20
indent-style = "space"
quote-style = "single"
